[#66] Compose code ignored.  - Compose related code moved into a compose package to be ignored by code coverage.
