+++ features/settings/presentation/src/main/java/com/feragusper/smokeanalytics/features/settings/presentation/SettingsViewModel.kt
@@ -3,7 +3,7 @@ package com.feragusper.smokeanalytics.features.settings.presentation
 import com.feragusper.smokeanalytics.features.settings.presentation.navigation.SettingsNavigator
 import com.feragusper.smokeanalytics.features.settings.presentation.mvi.SettingsIntent
 import com.feragusper.smokeanalytics.features.settings.presentation.mvi.SettingsResult
-import com.feragusper.smokeanalytics.features.settings.presentation.mvi.SettingsViewState
+import com.feragusper.smokeanalytics.features.settings.presentation.mvi.compose.SettingsViewState
 import com.feragusper.smokeanalytics.features.settings.presentation.process.SettingsProcessHolder
 import com.feragusper.smokeanalytics.libraries.architecture.presentation.MVIViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
