+++ features/settings/presentation/src/main/java/com/feragusper/smokeanalytics/features/settings/presentation/mvi/SettingsViewState.kt
@@ -1,142 +1,9 @@
 package com.feragusper.smokeanalytics.features.settings.presentation.mvi
 
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Button
-import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.Divider
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.SnackbarHost
-import androidx.compose.material3.SnackbarHostState
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.platform.testTag
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.dp
-import com.feragusper.smokeanalytics.features.settings.presentation.R
-import com.feragusper.smokeanalytics.libraries.architecture.presentation.extensions.versionName
-import com.feragusper.smokeanalytics.libraries.architecture.presentation.mvi.MVIViewState
-import com.feragusper.smokeanalytics.libraries.authentication.presentation.GoogleSignInComponent
-import com.feragusper.smokeanalytics.libraries.design.CombinedPreviews
-import com.feragusper.smokeanalytics.libraries.design.theme.SmokeAnalyticsTheme
-import kotlinx.coroutines.launch
-
-data class SettingsViewState(
-    internal val displayLoading: Boolean = false,
-    internal val currentEmail: String? = null,
-) : MVIViewState<SettingsIntent> {
-
-    interface TestTags {
-        companion object {
-            const val BUTTON_SIGN_OUT = "buttonSignOut"
-            const val BUTTON_SIGN_IN = "buttonSignIn"
-            const val VIEW_PROGRESS = "viewProgress"
-        }
-    }
-
-    @Composable
-    override fun Compose(intent: (SettingsIntent) -> Unit) {
-        val snackbarHostState = remember { SnackbarHostState() }
-        Scaffold(
-            snackbarHost = { SnackbarHost(snackbarHostState) }
-        ) {
-            Column(
-                modifier = Modifier
-                    .padding(it)
-                    .padding(vertical = 16.dp)
-                    .fillMaxSize(),
-                verticalArrangement = Arrangement.spacedBy(16.dp),
-            ) {
-                if (displayLoading) {
-                    Box(
-                        modifier = Modifier.fillMaxSize(),
-                        contentAlignment = Alignment.Center
-                    ) {
-                        CircularProgressIndicator(
-                            modifier = Modifier.testTag(TestTags.VIEW_PROGRESS)
-                        )
-                    }
-                } else {
-                    Text(
-                        modifier = Modifier.padding(16.dp),
-                        text = stringResource(R.string.settings_title),
-                        style = MaterialTheme.typography.titleLarge,
-                    )
-                    currentEmail?.let { email ->
-                        Row(
-                            modifier = Modifier.padding(horizontal = 16.dp),
-                            verticalAlignment = Alignment.CenterVertically,
-                        ) {
-                            Icon(
-                                painter = painterResource(id = R.drawable.ic_person),
-                                contentDescription = null
-                            )
-                            Column(modifier = Modifier.padding(horizontal = 16.dp)) {
-                                Text(
-                                    text = stringResource(id = R.string.settings_email),
-                                    style = MaterialTheme.typography.bodyLarge,
-                                    color = MaterialTheme.colorScheme.onSurface
-                                )
-                                Text(
-                                    text = email,
-                                    style = MaterialTheme.typography.bodyMedium,
-                                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                                )
-                            }
-                        }
-                        Divider()
-                        Button(
-                            modifier = Modifier
-                                .padding(horizontal = 16.dp)
-                                .testTag(TestTags.BUTTON_SIGN_OUT),
-                            onClick = { intent(SettingsIntent.SignOut) }
-                        ) {
-                            Text(text = stringResource(id = R.string.settings_logout))
-                        }
-                    } ?: run {
-                        val scope = rememberCoroutineScope()
-                        val context = LocalContext.current
-                        GoogleSignInComponent(
-                            modifier = Modifier.testTag(TestTags.BUTTON_SIGN_IN),
-                            onSignInSuccess = { intent(SettingsIntent.FetchUser) },
-                            onSignInError = {
-                                scope.launch {
-                                    snackbarHostState.showSnackbar(context.getString(com.feragusper.smokeanalytics.libraries.design.R.string.error_general))
-                                }
-                            },
-                        )
-
-                    }
-                    Spacer(modifier = Modifier.weight(1F))
-                    LocalContext.current.versionName()?.let { versionName ->
-                        Text(
-                            modifier = Modifier.fillMaxWidth(),
-                            text = versionName,
-                            style = MaterialTheme.typography.bodySmall,
-                            color = MaterialTheme.colorScheme.onSurface,
-                            textAlign = TextAlign.Center
-                        )
-                    }
-                }
-            }
-        }
-    }
-}
+import com.feragusper.smokeanalytics.features.settings.presentation.mvi.compose.SettingsViewState
+import com.feragusper.smokeanalytics.libraries.design.compose.CombinedPreviews
+import com.feragusper.smokeanalytics.libraries.design.compose.theme.SmokeAnalyticsTheme
 
 @CombinedPreviews
 @Composable
