+++ features/settings/presentation/src/main/java/com/feragusper/smokeanalytics/features/settings/presentation/mvi/compose/SettingsViewState.kt
@@ -1,4 +1,4 @@
-package com.feragusper.smokeanalytics.features.devtools.presentation.mvi
+package com.feragusper.smokeanalytics.features.settings.presentation.mvi.compose
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
@@ -8,6 +8,7 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Button
 import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
@@ -18,28 +19,37 @@ import androidx.compose.material3.SnackbarHostState
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
-import com.feragusper.smokeanalytics.features.devtools.presentation.R
+import com.feragusper.smokeanalytics.features.settings.presentation.R
+import com.feragusper.smokeanalytics.features.settings.presentation.mvi.SettingsIntent
 import com.feragusper.smokeanalytics.libraries.architecture.presentation.extensions.versionName
 import com.feragusper.smokeanalytics.libraries.architecture.presentation.mvi.MVIViewState
-import com.feragusper.smokeanalytics.libraries.design.compose.CombinedPreviews
-import com.feragusper.smokeanalytics.libraries.design.compose.theme.SmokeAnalyticsTheme
+import com.feragusper.smokeanalytics.libraries.authentication.presentation.compose.GoogleSignInComponent
+import kotlinx.coroutines.launch
 
-data class DevToolsViewState(
+data class SettingsViewState(
     internal val displayLoading: Boolean = false,
-    internal val currentUser: User? = null,
-) : MVIViewState<DevToolsIntent> {
+    internal val currentEmail: String? = null,
+) : MVIViewState<SettingsIntent> {
 
-    data class User(val email: String?, val id: String)
+    interface TestTags {
+        companion object {
+            const val BUTTON_SIGN_OUT = "buttonSignOut"
+            const val BUTTON_SIGN_IN = "buttonSignIn"
+            const val VIEW_PROGRESS = "viewProgress"
+        }
+    }
 
     @Composable
-    override fun Compose(intent: (DevToolsIntent) -> Unit) {
+    override fun Compose(intent: (SettingsIntent) -> Unit) {
         val snackbarHostState = remember { SnackbarHostState() }
         Scaffold(
             snackbarHost = { SnackbarHost(snackbarHostState) }
@@ -56,15 +66,17 @@ data class DevToolsViewState(
                         modifier = Modifier.fillMaxSize(),
                         contentAlignment = Alignment.Center
                     ) {
-                        CircularProgressIndicator()
+                        CircularProgressIndicator(
+                            modifier = Modifier.testTag(TestTags.VIEW_PROGRESS)
+                        )
                     }
                 } else {
                     Text(
                         modifier = Modifier.padding(16.dp),
-                        text = stringResource(R.string.devtools_title),
+                        text = stringResource(R.string.settings_title),
                         style = MaterialTheme.typography.titleLarge,
                     )
-                    currentUser?.let { user ->
+                    currentEmail?.let { email ->
                         Row(
                             modifier = Modifier.padding(horizontal = 16.dp),
                             verticalAlignment = Alignment.CenterVertically,
@@ -74,38 +86,40 @@ data class DevToolsViewState(
                                 contentDescription = null
                             )
                             Column(modifier = Modifier.padding(horizontal = 16.dp)) {
-                                user.email?.let { email ->
-                                    Text(
-                                        text = stringResource(id = R.string.devtools_email),
-                                        style = MaterialTheme.typography.bodyLarge,
-                                        color = MaterialTheme.colorScheme.onSurface
-                                    )
-                                    Text(
-                                        text = email,
-                                        style = MaterialTheme.typography.bodyMedium,
-                                        color = MaterialTheme.colorScheme.onSurfaceVariant,
-                                    )
-                                }
                                 Text(
-                                    text = stringResource(id = R.string.devtools_id),
+                                    text = stringResource(id = R.string.settings_email),
                                     style = MaterialTheme.typography.bodyLarge,
                                     color = MaterialTheme.colorScheme.onSurface
                                 )
                                 Text(
-                                    text = user.id,
+                                    text = email,
                                     style = MaterialTheme.typography.bodyMedium,
                                     color = MaterialTheme.colorScheme.onSurfaceVariant,
                                 )
                             }
                         }
                         HorizontalDivider()
+                        Button(
+                            modifier = Modifier
+                                .padding(horizontal = 16.dp)
+                                .testTag(TestTags.BUTTON_SIGN_OUT),
+                            onClick = { intent(SettingsIntent.SignOut) }
+                        ) {
+                            Text(text = stringResource(id = R.string.settings_logout))
+                        }
                     } ?: run {
-                        Text(
-                            modifier = Modifier.padding(16.dp),
-                            text = stringResource(id = R.string.devtools_logged_out),
-                            style = MaterialTheme.typography.bodyMedium,
-                            color = MaterialTheme.colorScheme.onSurfaceVariant,
+                        val scope = rememberCoroutineScope()
+                        val context = LocalContext.current
+                        GoogleSignInComponent(
+                            modifier = Modifier.testTag(TestTags.BUTTON_SIGN_IN),
+                            onSignInSuccess = { intent(SettingsIntent.FetchUser) },
+                            onSignInError = {
+                                scope.launch {
+                                    snackbarHostState.showSnackbar(context.getString(com.feragusper.smokeanalytics.libraries.design.R.string.error_general))
+                                }
+                            },
                         )
+
                     }
                     Spacer(modifier = Modifier.weight(1F))
                     LocalContext.current.versionName()?.let { versionName ->
@@ -120,36 +134,5 @@ data class DevToolsViewState(
                 }
             }
         }
-    }
-}
-
-@CombinedPreviews
-@Composable
-private fun DevToolsLoadingPreview() {
-    SmokeAnalyticsTheme {
-        DevToolsViewState(
-            displayLoading = true
-        ).Compose {}
-    }
-}
-
-@CombinedPreviews
-@Composable
-private fun DevToolsLoggedOutViewPreview() {
-    SmokeAnalyticsTheme {
-        DevToolsViewState().Compose {}
-    }
-}
-
-@CombinedPreviews
-@Composable
-private fun DevToolsLoggedInViewPreview() {
-    SmokeAnalyticsTheme {
-        DevToolsViewState(
-            currentUser = DevToolsViewState.User(
-                id = "123",
-                email = "fernancho@gmail.com"
-            )
-        ).Compose {}
     }
 }
