+++ features/stats/presentation/src/main/java/com/feragusper/smokeanalytics/features/stats/presentation/StatsViewModel.kt
@@ -3,7 +3,7 @@ package com.feragusper.smokeanalytics.features.stats.presentation
 import com.feragusper.smokeanalytics.features.stats.navigation.StatsNavigator
 import com.feragusper.smokeanalytics.features.stats.presentation.mvi.StatsIntent
 import com.feragusper.smokeanalytics.features.stats.presentation.mvi.StatsResult
-import com.feragusper.smokeanalytics.features.stats.presentation.mvi.StatsViewState
+import com.feragusper.smokeanalytics.features.stats.presentation.mvi.compose.StatsViewState
 import com.feragusper.smokeanalytics.features.stats.presentation.process.StatsProcessHolder
 import com.feragusper.smokeanalytics.libraries.architecture.presentation.MVIViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
