+++ features/stats/presentation/src/main/java/com/feragusper/smokeanalytics/features/stats/presentation/mvi/compose/StatsViewState.kt
@@ -1,0 +1,61 @@
+package com.feragusper.smokeanalytics.features.stats.presentation.mvi.compose
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.SnackbarHost
+import androidx.compose.material3.SnackbarHostState
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import com.feragusper.smokeanalytics.features.stats.presentation.R
+import com.feragusper.smokeanalytics.features.stats.presentation.mvi.StatsIntent
+import com.feragusper.smokeanalytics.libraries.architecture.presentation.mvi.MVIViewState
+
+class StatsViewState : MVIViewState<StatsIntent> {
+
+    @Composable
+    override fun Compose(intent: (StatsIntent) -> Unit) {
+        val snackbarHostState = remember { SnackbarHostState() }
+        Scaffold(
+            snackbarHost = { SnackbarHost(snackbarHostState) }
+        ) {
+            Column(
+                modifier = Modifier
+                    .padding(it)
+                    .padding(vertical = 16.dp)
+                    .fillMaxSize(),
+                verticalArrangement = Arrangement.Center,
+                horizontalAlignment = Alignment.CenterHorizontally
+            ) {
+                Image(
+                    painter = painterResource(id = R.drawable.ill_chart),
+                    contentDescription = null
+                )
+                Text(
+                    modifier = Modifier.padding(top = 24.dp),
+                    text = stringResource(R.string.stats_title),
+                    style = MaterialTheme.typography.titleLarge,
+                )
+                Text(
+                    modifier = Modifier
+                        .padding(top = 8.dp)
+                        .padding(horizontal = 31.dp),
+                    text = stringResource(R.string.stats_subtitle),
+                    style = MaterialTheme.typography.bodyMedium,
+                    textAlign = TextAlign.Center
+                )
+            }
+        }
+    }
+}
