+++ libraries/design/src/main/java/com/feragusper/smokeanalytics/libraries/design/compose/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.feragusper.smokeanalytics.libraries.design.theme
+package com.feragusper.smokeanalytics.libraries.design.compose.theme
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
@@ -18,8 +18,8 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
-import com.feragusper.smokeanalytics.libraries.design.ColorsPreview
-import com.feragusper.smokeanalytics.libraries.design.UiModeThemePreview
+import com.feragusper.smokeanalytics.libraries.design.compose.ColorsPreview
+import com.feragusper.smokeanalytics.libraries.design.compose.UiModeThemePreview
 
 internal object PaletteTokens {
